use regular timestamp type
also add missing documentation comments

+++ evm-contracts/src/v0.6/dev/ConversionProxy.sol
@@ -109,6 +109,26 @@ contract ConversionProxy is AggregatorInterface, Owned {
     return _getTimestamp(_roundId);
   }
 
+  /**
+   * @notice get all details about a round. Consumers are encouraged to check
+   * that they're receiving fresh data by inspecting the updatedAt and
+   * answeredInRound return values.
+   * @param _roundId the round ID to retrieve the details for. If _roundId
+   * has the special value UINT256_MAX (2**256-1), the contract will retrieve
+   * the latest round's details.
+   * @return roundId is the round ID for which details were retrieved
+   * @return answer is the answer for the given round
+   * @return startedAt is the timestamp when the round was started. This is 0
+   * if the round hasn't been started yet.
+   * @return updatedAt is the timestamp when the round last was updated (i.e.
+   * answer was last computed)
+   * @return answeredInRound is the round ID of the round in which the answer
+   * was computed. answeredInRound may be smaller than roundId when the round
+   * timed out. answerInRound is equal to roundId when the round didn't time out
+   * and was completed regularly.
+   * @dev Note that for in-progress rounds (i.e. rounds that haven't yet received
+   * maxSubmissions) answer and updatedAt may change between queries.
+   */
   function getRoundData(uint256 _roundId)
     external
     view
@@ -117,8 +137,8 @@ contract ConversionProxy is AggregatorInterface, Owned {
     returns (
       uint256 roundId,
       int256 answer,
-      uint64 startedAt,
-      uint64 updatedAt,
+      uint256 startedAt,
+      uint256 updatedAt,
       uint256 answeredInRound
     )
   {
@@ -148,8 +168,8 @@ contract ConversionProxy is AggregatorInterface, Owned {
     returns (
       uint256 roundId,
       int256 answer,
-      uint64 startedAt,
-      uint64 updatedAt,
+      uint256 startedAt,
+      uint256 updatedAt,
       uint256 answeredInRound
     )
   {
@@ -191,15 +211,15 @@ contract ConversionProxy is AggregatorInterface, Owned {
     returns (
       uint256 roundId,
       int256 answer,
-      uint64 startedAt,
-      uint64 updatedAt,
+      uint256 startedAt,
+      uint256 updatedAt,
       uint256 answeredInRound
     )
   {
     uint256 roundIdFrom;
     int256 answerFrom;
-    uint64 startedAtFrom;
-    uint64 updatedAtFrom;
+    uint256 startedAtFrom;
+    uint256 updatedAtFrom;
     uint256 answeredInRoundFrom;
 
     (roundIdFrom, answerFrom, startedAtFrom, updatedAtFrom, answeredInRoundFrom) = from.getRoundData(_roundId);
@@ -212,15 +232,15 @@ contract ConversionProxy is AggregatorInterface, Owned {
     returns (
       uint256 roundId,
       int256 answer,
-      uint64 startedAt,
-      uint64 updatedAt,
+      uint256 startedAt,
+      uint256 updatedAt,
       uint256 answeredInRound
     )
   {
     uint256 roundIdFrom;
     int256 answerFrom;
-    uint64 startedAtFrom;
-    uint64 updatedAtFrom;
+    uint256 startedAtFrom;
+    uint256 updatedAtFrom;
     uint256 answeredInRoundFrom;
 
     (roundIdFrom, answerFrom, startedAtFrom, updatedAtFrom, answeredInRoundFrom) = from.getRoundData(_latestRound());
